+++ src/components/Page/Single/schema.ts
