+++ src/components/Page/Single/types.ts
